MultiDoc-GroqQA/requirements.txt
streamlit
python-dotenv
PyPDF2
faiss-cpu
groq
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
sentence-transformers
